requirements.txt
python-igraph
setuptools